add-ext2-fs.patch
This adds ext2 support from U-Boot and stitches it into the
partition stuff.  It also upgrades the board definitions so they
can define the path to look for in the ext2 filesystem being
mounted.  I used /boot/uImage.bin because this is already in use
by the packaged kernel.

We now mount, open and pull the kernel from ext2 in phase2.c if
the kernel source defines it.

Signed-off-by: Andy Green <andy@openmoko.com>