add-openocd-script.patch
This is the script for gta03 load by openocd.  We use the single qi
image for all purposes now so U-Boot lowlevel_foo style crap is dead.

We have to run the first part of init before copying qi image to
TEXT_BASE (0x33000000) because DRAM is not functional until then.

Fixed branchthrough at 0x8 is used as a place to hang the breakpoint.

Signed-off-by: Andy Green <andy@openmoko.com>