qi-add-6410-bootable-sdcard-generator.patch
This patch adds a small script that knows how to partition, format and
prepare with a bootloader image an SD Card for usage with SD Card boot
on 6410.

You use it like this:

./6410-partition-sd.sh sde sdhc ./image/qi-s3c6410-andy_495294c60f2f1432

This will prepare your card with three partitions and put the bootloader
images at the end as required by 6410 iROM.

If you put anything on the fourth parameter, it skips the fdisk and the
formatting stages and just updates the bootloader.

Signed-off-by: Andy Green <andy@openmoko.com>