qi-introduce-device-named-kernels.patch
This patch changes Qi to insist to fetch device-specific kernels from
filesystems that have named kernels.  The kernels looked for are now
called

GTA01:  /boot/uImage-GTA01.bin
GTA02:  /boot/uImage-GTA02.bin
GTA03:  /boot/uImage-GTA03.bin

This is part of the support for single rootfs that can be run on
multiple devices with correct kernel and module handling by the
bootloader.

Signed-off-by: Andy Green <andy@openmoko.com>