qi-gta02-no-inidcators-if-battery-low.patch
This patch turns on the SYS and BAT monitoring filters, and checks if the battery
meets the BAT OK threshold.  If it doesn't, which is the case if the battery is
not present, it disables the "indicator" (eg, LED, vibrator) stuff and holds the
CPU at 200MHz during the boot into Linux.

This allows the GTA02 A6 here to boot with no battery up to the point it is
going to bring up backlight, and this with 100mA limit on USB at PMU.  Enabling
the threshold filters for battery and SYS seems to have been critical in
getting any stability with this.

Signed-off-by: Andy Green <andy@openmoko.com>