phase2: simplify the bootloader_second_phase mega-loop
This puts the loop body in a new function called try_this_kernel.
As an added benefit, we can drop one level of indentation.

This change is hard to read as a patch. It gets better if one just
applies it and then looks at it with "git diff -w" or similar.

Signed-off-by: Werner Almesberger <werner@openmoko.org>