qi-fix-block-init-cache-logic.patch
Changes in the last couple of weeks aimed at cleaning this
code broke the block device init cache logic.  This patch
restores the logic and reduces the card init failures when
there is no SD Card present to 1 regardless of the number
of partitions probed on the card.

Together with the reduction in Glamo card wait on init this
reduces the delay before trying NAND to 1/9th of before the
patches.

Signed-off-by: Andy Green <andy@openmoko.com>