My revised version of the partition tool
The fdisk and mkfs was factored out into a shell function
so that it is easier to create dirfferent partition layouts.

The default case was left the same as the original script

The partitions can be specified as percentages of the
available space (after Qi reserved space has been taken out)
by providing a comma separated list of values:
e.g. 25,0,50
Which would create partitions 1 and 3 with 25% and 50%
of the usable capacity. Partitions 2 and 4 would not be
assigned.

Signed-off-by: Christopher Hall <hsw@openmoko.com>