Bad magic when booting from NAND
Qi fails to boot from NAND. Qi booting from SD goes fine. Booting NAND
from NOR works as well.

I have found the problem. The error was because we check if the
current block is bad or the next one is bad. And skip the current
block even if only the next one is bad. This way we skip good block
before a bad one.

This patch fixes the bug.

Signed-off-by: Dmitry Kurochkin <dmitry.kurochkin@gmail.com>